requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alabaster=0.7.16=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
babel=2.14.0=pypi_0
beartype=0.17.2=pypi_0
black=24.2.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=pypi_0
cfgv=3.4.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
coverage=7.4.1=pypi_0
distlib=0.3.8=pypi_0
docutils=0.20.1=pypi_0
filelock=3.13.1=pypi_0
flake8=7.0.0=pypi_0
gtsam=4.2=pypi_0
hydra-core=1.3.2=pypi_0
identify=2.5.34=pypi_0
idna=3.6=pypi_0
imagesize=1.4.1=pypi_0
iniconfig=2.0.0=pypi_0
isort=5.13.2=pypi_0
jinja2=3.1.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lz4=4.3.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mccabe=0.7.0=pypi_0
mdurl=0.1.2=pypi_0
mypy=1.8.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nodeenv=1.8.0=pypi_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pypi_0
pathspec=0.12.1=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py311h06a4308_0
platformdirs=4.2.0=pypi_0
pluggy=1.4.0=pypi_0
plum-dispatch=2.3.2=pypi_0
pre-commit=3.6.1=pypi_0
psutil=5.9.8=pypi_0
pycodestyle=2.11.1=pypi_0
pyflakes=3.2.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pytest=8.0.1=pypi_0
python=3.11.7=h955ad1f_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rosbags=0.9.19=pypi_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
setuptools=69.1.0=pypi_0
snowballstemmer=2.2.0=pypi_0
sphinx=7.2.6=pypi_0
sphinxcontrib-applehelp=1.0.8=pypi_0
sphinxcontrib-devhelp=1.0.6=pypi_0
sphinxcontrib-htmlhelp=2.0.5=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.7=pypi_0
sphinxcontrib-serializinghtml=1.1.10=pypi_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.2=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023d=h04d1e81_0
urllib3=2.2.0=pypi_0
virtualenv=20.25.0=pypi_0
wheel=0.41.2=py311h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstandard=0.22.0=pypi_0
